requirements.txt
langchain
langchain-community
langchain-experimental
fastapi
python-dotenv
streamlit
uvicorn
pydantic
httpx
requests
langchain_google_community
langchain_google_genai
langchain_tavily
langchain_groq
langchain_openai
langgraph


-e .
